requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
flask-cors==4.0.0
python-dotenv==1.0.0
flask-bcrypt==1.0.1
flask-login==0.6.3
flask-wtf==1.2.2
email-validator==2.3.0